Refactor "actions" inside module
Made it so they don't rely on the module state so they could be split
out into a separate module if I wanted. Not sure if this will introduce
subtle bugs but hopefully when I show this to people someone can correct
me if it does and then I'm screwing up quickly.