Added a basic test harness with Jasmine
At the moment I've just got a test there for a React component, and it's not
using Jest which might mean that testing some more complex stuff doesn't work
but it is working ok for basic stuff and the important bit I want to test I
guess is the RxJS stuff which I don't need React test tooling for. Run the
tests with `npm test` and you'll see a browser window pop up with the jasmine
test results, updating tests will reload the jasmine page and you need to add
your new spec files in test/js/webpack.config.js entry point. This was based
on https://egghead.io/lessons/react-unit-testing-a-react-component-using-jasmine-and-webpack